feat(frontend): iplemented redux injection in server side and wrapper for navar
